backend/requirements.txt
fastapi==0.110.0
strawberry-graphql[fastapi]==0.215.1
uvicorn==0.29.0
python-multipart>=0.0.5,<0.0.7
Pillow==10.3.0
protobuf==4.25.3
torch==2.5.1
torchvision==0.20.1
transformers==4.49.0
git+https://github.com/joanrod/star-vector.git
omegaconf==2.3.0
# flash_attn REMOVED, dont want to use GPU